// src/app/models/tender.ts
export class Tender {
     // tslint:disable-next-line: variable-name
     _id: number;
     tenderNumber: string;
     tenderDescription: string;
     category: string;
     tenderStatus: any;
     nameOfInstitution: any;
    //  options: any;
     InstitutionContactPerson: string;
     InstitutionPersonEmail: string;
     InstitutionPersonPhone: string;
     awardedPoint: number;
     datePublished: string;
     closingDate: string;
}